backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
boto3
pydantic
requests
python-multipart